The init of all structs dongles mutexses and cond is pretty much done now i need to improve free logic the covers every possibility. If something fails mid init. and if everything was successfull

+++ helper.c
@@ -7,7 +7,7 @@ int	ft_isdigit(int c)
 	return (0);
 }
 
-long	get_cur_time_ns(struct timeval tv)
+long long	get_cur_time_ns(struct timeval tv)
 {
 	long	time_ms;
